Test: 식재료 찜 목록

+++ back/src/main/java/com/gamul/api/request/IngredientSelectReq.java
@@ -15,5 +15,5 @@ public class IngredientSelectReq {
     String userName;
 
     @JsonProperty("ingredient_id")
-    int ingredientId;
+    Long ingredientId;
 }
